requirements.txt
python-dotenv==1.0.0
fastapi==0.103.1
python-multipart==0.0.6
pdf2image==1.16.3
pdfminer.six==20221105
PyMuPDF==1.23.6
openai==1.3.0
requests==2.31.0
streamlit==1.27.2
uvicorn==0.23.2